add docs

+++ build.sh
@@ -1,25 +1,28 @@
 #!/bin/bash
 
-rm -r target
-rm main.bc
 
-if [ ! -d "target" ]; then
-    mkdir target
+if [ -f "main.bc" ]; then
+    rm main.bc
 fi
+if [ -d "target" ]; then
+    rm -r target
+fi
+mkdir target
+
 if [ ! -d "reports" ]; then
     mkdir reports
 fi
 
-# # 遍历 src 目录下的所有.c文件
-# for file in src/*.c
-# do
-#     # compile to bc
-#     # echo $file
-#     # 获取 base name
-#     base=$(basename $file .c)
-#     # echo $base
-#     clang -I /snap/klee/9/usr/local/include -I ./headers -emit-llvm -c -g -O0 -Xclang -disable-O0-optnone $file -o target/$base.bc
-# done
+# 遍历 src 目录下的所有.cc文件
+for file in src/*.cc
+do
+    # compile to bc
+    # echo $file
+    # 获取 base name
+    base=$(basename $file .cc)
+    # echo $base
+    clang -I /snap/klee/9/usr/local/include -I ./headers -emit-llvm -c -g -O0 -Xclang -disable-O0-optnone $file -o target/$base.bc
+done
 
 # compile to bc
 clang -I /snap/klee/9/usr/local/include -I ./headers -emit-llvm -c -g -O0 -Xclang -disable-O0-optnone main.cc
@@ -49,6 +52,9 @@ for file in reports/klee-last/*.ktest
 do
     echo $file
     KTEST_FILE=$file $gener
-    echo $? # 打印返回值
+    ret=$? # 获取返回值
+    # 换行
+    echo
+    echo "return value: $ret"
 done
